% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antennaNeighbours.R
\name{antennaNeighbours}
\alias{antennaNeighbours}
\title{Computes a list of neighbouring antennas.}
\usage{
antennaNeighbours(coverarea)
}
\arguments{
\item{coverarea}{a data.table object with two columns: 'antennaID' and 'cell'. The first column contains the ID of
each antenna while the second one contains an \code{sp} object that represents the coverage area of the corresponding
antenna. It is obtained by calling \code{readCells()} function.}
}
\value{
A data.table object with a single column called 'nei'. Each element of this column is a pair of the form
  antennaID1-antennaID2 where the two antennas are considered neighbours.
}
\description{
Computes a list of pairs antennaID1-antennaID2 with neighbouring antennas. Two antennas are considered
  neighbours if their coverage areas overlap (i.e. their intersection is not void).
}
