% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aprioriDuplicityProb.R
\name{aprioriDuplicityProb}
\alias{aprioriDuplicityProb}
\title{Apriori probability of duplicity.}
\usage{
aprioriDuplicityProb(prob2Devices, ndevices)
}
\arguments{
\item{prob2Devices}{The probability of a person to have 2 devices. In case of using simulated data, this parameter is
read from the simulation configuration file.}

\item{ndevices}{The number of devices detected by the network during the time horizon under consideration.}
}
\value{
Apriori probability of duplicity, i.e. the probability of a device to be in a 2-to-1 correspondence with the
  person who holds it.
}
\description{
Apriori probability of duplicity, i.e. the probability of a device to be in a 2-to-1 correspondence with
  the person who holds it. It is computed as \eqn{ 2 * (Ndev-Ndev2) / (Ndev * (Ndev-1)) } where \eqn{Ndev} is the
  total number of devices and \eqn{Ndev2} is the number of devices that are in a 1-to-1 correspondence with the
  persons that hold them.
}
