% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCentroids.R
\name{buildCentroids}
\alias{buildCentroids}
\title{Builds the centroid of the rectangular tiles.}
\usage{
buildCentroids(ntiles_x, ntiles_y, tile_sizeX, tile_sizeY)
}
\arguments{
\item{ntiles_x}{The number of tiles on OX.}

\item{ntiles_y}{The number of tiles on OY.}
}
\value{
A data.table object with three columns: \code{tile ID, centroidCoord_x \cr 
centroidCoord_y}.
}
\description{
Builds the centroid of each (rectangular) tile in the grid. 
The centroid is the center of the tile, i.e. the intersection point 
of the diagonals of the tile.
}
\keyword{internal}
