% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDeltaProb.R
\name{buildDeltaProb}
\alias{buildDeltaProb}
\title{Computes the probability distribution for Delta X and Delta Y.}
\usage{
buildDeltaProb(xp1, xp2)
}
\arguments{
\item{xp1}{A data.table object with 3 columns representing the x and y coordinates of the centroids of tiles in the
grid and the location probability for each corresponding tile for the first device.}

\item{xp2}{a data.table object with 3 columns representing the x and y coordinates of the centroids of tiles in the
grid and the location probability for each corresponding tile for the second device.}
}
\value{
a list with two data.table objects, the first is the probability distribution for Delta X and the second for
  Delta Y. The elements are named DeltaX and DeltaY.
}
\description{
Computes the probability distribution for Delta X and Delta Y where Delta X and Delta Y are the
  differences between the centroids of the tiles on OX and OY. For a detailed explanation of what Delta X and Delta Y
  mean please reffer to
  \href{https://webgate.ec.europa.eu/fpfis/mwikis/essnetbigdata/images/f/fb/WPI_Deliverable_I3_A_proposed_production_framework_with_mobile_network_data_2020_05_31_draft.pdf}{WPI
   Deliverable 3} The posterior location probability for each tile and each device comes from \pkg{destim}.
}
