% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDeltaProb2.R
\name{buildDeltaProb2}
\alias{buildDeltaProb2}
\title{Computes the probability distribution for Delta X and Delta Y.}
\usage{
buildDeltaProb2(centroids, postLocProbDevice1, postLocProbDevice2)
}
\arguments{
\item{centroids}{The centroids of all tiles in the grid.}

\item{postLocProbDevice1}{The location probability of the first device, at a certain time instant, for each tile in
the grid.}

\item{postLocProbDevice2}{The location probability of the first device, at a certain time instant, for each tile in
the grid.}
}
\value{
a list with two data.table objects, the first is the probability distribution for Delta X and the second for
  Delta Y. The elements are named DeltaX and DeltaY.
}
\description{
Computes the probability distribution for Delta X and Delta Y where Delta X / Delta Y are the
  differences between the centroids of the tiles on OX / OY. For a detailed explanation of what Delta X and Delta Y
  mean please reffer to
  \href{https://webgate.ec.europa.eu/fpfis/mwikis/essnetbigdata/images/f/fb/WPI_Deliverable_I3_A_proposed_production_framework_with_mobile_network_data_2020_05_31_draft.pdf}{WPI
   Deliverable 3} The posterior location probability for each tile nd each device comes from \pkg{destim}.
}
