% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDuplicityTable1to1.R
\name{buildDuplicityTable1to1}
\alias{buildDuplicityTable1to1}
\title{Builds a data.table object with device IDs and the duplicity probability for each device.}
\usage{
buildDuplicityTable1to1(res, devices, Pii, lambda = NULL)
}
\arguments{
\item{res}{A list of matrix objects returned by each worker node with the number of rows corresponding to the number
of devices allocated to an working node and the number of columns equals to the number of devices.}

\item{devices}{A vector with device IDs.}

\item{Pii}{Apriori probability of devices to be in 1-to-1 correspondence with the owner.}

\item{lambda}{If this parameter is non NULL, the duplicity probabilities are computed according to the approach described in
"An end-to-end statistical process with mobile network data for Official Statistics" paper.}
}
\value{
A data.table object with two columns: the device IDs and the corresponding duplicity probability for each
  device.
}
\description{
Builds a data.table object with with two columns: device IDs and the duplicity probability for each
  device. This function is called in case of using the "1to1" method of computing the duplicity probabilities,
  receiving a list of matrix objects returned by each worker node with lines corresponding to a subset of devices.
  These parts a put together and form a symmetic matrix used to compute the duplicity probability. It is a utility
  function and it is not accesible from outside the package.
}
