% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerOfProbabilities.R
\name{centerOfProbabilities}
\alias{centerOfProbabilities}
\title{Computes the center of probabilities (as center of mass) for each posterior probability cloud, at each time
  instant, for a device.}
\usage{
centerOfProbabilities(centroid, postLocProb)
}
\arguments{
\item{centroid}{A data.table object with the centroids for each tile in the grid.}

\item{postLocProb}{a Matrix object with the posterior location probabilities computed by \code{destim} for a device.
The number of rows equals the number of tiles and the number of columns equals the number of successive time
instants. An element \code{postLocProb[i,j]} represents the posterior location probability corresponding to tile
\code{i} at the time instant \code{j}.}
}
\value{
A table with two columns: centerProb_x and centerProb_y representing the x and y coordinates of the centers
  of probabilities.
}
\description{
Computes the center of probabilities (as center of mass) for each posterior probability cloud, at each
  time instant, for a device. It uses the results obtained from \pkg{destim} package.
}
