% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConnections_step2.R
\name{checkConnections_step2}
\alias{checkConnections_step2}
\title{Make the second step of the checks between the connections and the emission model.}
\usage{
checkConnections_step2(emissionProbs, connections, gridParams)
}
\arguments{
\item{emissionProbs}{the matrix of emission probabilities.}

\item{connections}{a matrix with the connections, a row per device and
a column per time.}

\item{gridParams}{a list with the parameters of the grid: nrow, ncol, tileX, tileY.}
}
\value{
A \code{list} with two elements, one is the information about the checks
with a vector containing all the time padding coefficients and the second 
is the matrix of connections with the time padding done by using the maximum
coefficient.
}
\description{
This function obtains the results of the second step of checks
for the compatibility between the connections observed and the emission model. 
It checks that the transitions from one connections to the following is
coherent with the probabilities in the emission matrix. In those cases when 
the transition is not compatible, a solution to fit the model is given 
with the needed time padding coefficient.
}
\examples{
\dontrun{
}

}
