% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDuplicityBayesian.R
\name{computeDuplicityBayesian}
\alias{computeDuplicityBayesian}
\title{Computes the duplicity probabilities for each device using a Bayesian approach.}
\usage{
computeDuplicityBayesian(
  method,
  deviceIDs,
  pairs4dupl,
  modeljoin,
  llik,
  P1 = NULL,
  Pii = NULL,
  init = TRUE
)
}
\arguments{
\item{deviceIDs}{A vector with the all device IDs. It is obtained by caliing the \code{getDevices()} function.}

\item{pairs4dupl}{A data.table object with pairs of devices and pairs of antennas where these devices are connecting.
It can be obtained by calling \code{computePairs()} function.}

\item{modeljoin}{The joint HMM model returned by \code{getJointModel()} function.}

\item{llik}{A vector with the values of the log likelihood after the individual HMM models for each device were
fitted. This vector can be obtained by calling \code{fitModels()} function.}

\item{P1}{The apriori duplicity probility as it is returned by \code{aprioriDuplicityProb()} function. It is used
when "pairs" method is selected.}

\item{Pii}{Apriori probability of a device to be in a 1-to-1 correspondence with the holder as it is returned by
\code{aprioriOneDeviceProb()} function. This parameter is used only when "1to1" method is selected.}

\item{init}{A logical value. If TRUE, the \code{fit()} function uses the stored steady state as fixed initialization,
otherwise the steady state is computed at every call of \code{fit()} function.}

\item{method.}{Selects a method to compute the duplicity probabilities. It could have one of the two values: "pairs"
or "1to1". When selecting "pairs" method, the pairs4dupl parameter contains only the compatible pairs of devices,
i.e. the pairs that most of the time are connected to the same or to neighbouring antennas. "1to1" method checks
all posible combinations between devices to compute the duplicity probabilities.}
}
\value{
a data.table object with two columns: 'deviceID' and 'dupP'. On the first column there are deviceIDs and on
  the second column the corresponding duplicity probability, i.e. the probability that a device is in a 2-to-1
  correspondence with the holder.
}
\description{
Computes the duplicity probabilities for each device using a Bayesian approach.
}
