% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersionRadius.R
\name{dispersionRadius}
\alias{dispersionRadius}
\title{Computes the radius of dispersion for the probability cloud of a device at each time
  instant.}
\usage{
dispersionRadius(
  centroid,
  postLocProb,
  center = NULL,
  method = "euclidean",
  p = 2
)
}
\arguments{
\item{centroid}{The centroids of all tiles in the grid.}

\item{center}{The center of posterior location probabilities. If it is not provided, it is
computed here.}

\item{method}{The distance to be used. Currently only the euclidean distance is implemented.}

\item{p}{the power of the Minkowski distance. The default value is 2 (euclidean distance).}
}
\value{

}
\description{
Computes the radius of dispersion for the probability cloud of a device at each
  time instant. The posterior location probabilities for a specific device, for each tile and
  each time instant, is provided as a parameter.
}
