% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example1.R
\name{example1}
\alias{example1}
\title{Example of using deduplication package - the simple way}
\usage{
example1()
}
\description{
#This is just an example on how to compute duplicity probabilities using simulated data. All the files used in this
#example are supposed to be produced using the simulation software. The "simulation.xml" file is an exeception and it
#is an input file for the simulation software. The files used in this example are provided with the deduplication
#package.
}
