% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModels.R
\name{fitModels}
\alias{fitModels}
\title{Fits the HMM models for each device.}
\usage{
fitModels(ndevices, model, connections, parallel = TRUE)
}
\arguments{
\item{ndevices}{The number of devices.}
}
\value{

}
\description{
Fits the HMM model for each device usin the \code{fit()} function from \code{destim} package.
}
