% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeviceIDs.R
\name{getDeviceIDs}
\alias{getDeviceIDs}
\title{Builds a vector with the IDs of the mobile devices.}
\usage{
getDeviceIDs(events)
}
\arguments{
\item{events}{A data.table object that contains the events generated by the mobile network. It is returned by the
\code{readEvents()} function. The device IDs are on the second column of this data.table object.}
}
\value{
A vector with the IDs of the mobile devices detected by the network.
}
\description{
Builds a vector with the IDs of the mobile devices by taking the unique values from the events data set.
}
