% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmissionProbs.R
\name{getEmissionProbs}
\alias{getEmissionProbs}
\title{Builds the emission probabilities for the HMM used to estimate the 
posterior location probabilities.}
\usage{
getEmissionProbs(
  nrows,
  ncols,
  signalFileName,
  sigMin,
  handoverType = "strength",
  simulatedData = TRUE,
  emissionModel = NULL,
  antennaFileName = NULL
)
}
\arguments{
\item{signalFileName}{The name of the .csv file that contains the signal 
strength/quality for each tile in the grid. This file is one of the outputs of 
the data simulator. The data are organized as a matrix with the number of rows
equals to the number of antennas and the the following columns: \cr 
\code{Antenna ID, Tile 0, Tile 1, ... Tile (N-1)}.
On the first column there are the antenna IDs and on the rest of the columns 
the corresponding signal strength/quality for each tile in the grid.}

\item{sigMin}{The minimum value of the signal strength/quality that allow a 
connection between a device and an antenna.}

\item{handoverType}{The handover mechanism used by the mobile network. It could 
have two values: "strength" or "quality". It should match the types of the 
values in the signal file, otherwise the results are impredictible.}

\item{simulatedData}{If TRUE, the input data provided to this function come 
from the simulator otherwise the data come from a real mobile network.}

\item{emissionModel}{A parameter that can take two values: "RSS" or "SDM". 
It indicates how the emission probabilities are computed. This parameter is 
needed to force computing the emission probabilities with a "wrong" model. For
example, the signal file contains the values of the signal strength, the 
\code{handoverType} parameter is set to 'strength' but the \code{emissionModel} 
is set to "SDM", the values of the signal strength are transformed in signal 
quality and the emission probabilities are computed using the signal quality. 
Such a combination should never be used in practice but it is allowed only for 
demonstrative purposes: it can be use to demonstrate that if the emission 
probabilities are not correctly computed then the resulted duplicity 
probabilities are wrong.}

\item{antennaFileName}{This parameter is needed to read the technical 
parameters of antennas.These parameters are used to transform the signal 
strength in signal quality and the other way around. They are needed only in 
the case the emission probabilities are computed using the signal quality when 
the \code{handoverType} is "strength" or when they are computed
 using signal quality when the \code{handoverType} is "quality".}

\item{nrow}{the number of rows in the grid. It can be obtained by calling 
\code{readGridParams()}.}

\item{ncol}{the number of columns in the grid. It can be obtained by calling 
\code{readGridParams()}.}
}
\value{
Returns a Matrix object with the emission probabilities for the HMM. 
The number of rows equals the number of tiles in the grid and the number of 
columns equals the number of antennas. An element (i,j) of this matrix
corresponds to the probability of a device being in tile i to be connected to 
antenna j. The row names of the matrix are the tile indexes and the column 
names are the antenna IDs.
}
\description{
Builds the emission probabilities needed for the HMM used to 
estimate the posterior location probabilities. In case of using simulated data, 
these probabilities are build using the signal strength or signal
 quality saved by the simulation software for each tile in the grid.
}
