% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmissionProbsJointModel.R
\name{getEmissionProbsJointModel}
\alias{getEmissionProbsJointModel}
\title{Builds the emission probabilities for the joint HMM.}
\usage{
getEmissionProbsJointModel(emissionProbs)
}
\arguments{
\item{emissionProbs}{the emission probabilities (the location probabilities) computed by calling
\code{getEmissionProbs()} for each individual device.}
}
\value{
Returns a matrix with the joint emission probabilities for the HMM. The number of rows equals the number
  tiles and the number of columns equals the number of combinations between antenna IDs. Before the combination
  between antenna IDs are build, the NA value is added to the list of antenna IDs. An element in this matrix
  represents the transition probability from an antenna to another, computed for each tile in the grid.
}
\description{
Builds the emissions probabilities needed for the joint HMM used to estimate the posterior location
  probabilitities.
}
