% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenericModel.R
\name{getGenericModel}
\alias{getGenericModel}
\title{Buildsw the generic HMM model.}
\usage{
getGenericModel(nrows, ncols, emmissionProbs, initSteady = TRUE)
}
\arguments{
\item{nrows}{Number of rows in the grid.}

\item{ncols}{Number of columns in the grid.}

\item{initSteady}{If true the initial a-priori distribution is set to the steady state of the transition matrix.}

\item{emmission}{probs A matrix with the event location probabilities. The number of rows equals the numberof
tiles in the grid and the number of columns equals the number of antennas plus one.}
}
\value{
Returns an HMM model with the initial a-priori distribution set to the steady state of the transition matrix.
}
\description{
Buildsw the generic HMM model using the emmission probabilities which tare the event locations
computed using the mobile network parameters.
}
