% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEvents.R
\name{readEvents}
\alias{readEvents}
\title{Reads the network events file.}
\usage{
readEvents(eventsFileName, simulatedData = TRUE)
}
\arguments{
\item{eventsFileName}{The name of the file with the network events to be used. Depending on the parameter
\code{simulatedData} it could be a .csv file coming from the simulation software or from a real MNO. In case the
file comes from the simulation software it should contain following columns: \code{time, antennaID, eventCode,
deviceID, x, y, tile}. Only the first 4 columns are used, the rest are ignored.}

\item{simulatedData}{If TRUE it means that the input data are simulated data, otherwise the data come from a real
MNO.}
}
\value{
Returns a data.table object that contains the events generated by the mobile network. The number of rows
  equals the number of connection events recorded by the network. The returned object has the following columns:
  \code{time, deviceID, eventCode, antennaID, x, y, tile, obsVar}. \code{obsVar} stands for observed variable and is
  a concatenation between the antenna ID and the event code.
}
\description{
Reads the network events file. This file can come from the network simulator or it ca be a file with
  real mobile network events provided by an MNO.
}
