% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPostLocJointProb.R
\name{readPostLocJointProb}
\alias{readPostLocJointProb}
\title{Read a file with the posterior location joint probabilities.}
\usage{
readPostLocJointProb(path, prefixName, deviceID)
}
\arguments{
\item{path}{The path to the location where the posterior location probabilities are stored. The file with the
location probabilities should have the name \code{postLocDevice_ID.csv} where \code{ID} is replaced with the device
ID.}

\item{prefixName}{The file name prefix. The whole file name is composed by a concatenation of \code{prefixName},
\code{_} and \code{deviceID}.}

\item{deviceID}{The device ID for which the posterior location probabilities are read.}
}
\value{
A Matrix object with the posterior location probabilities for the device with ID equals to deviceID. A row
  corresponds to a tile and a column corresponds to a time instant.
}
\description{
Read a csv file with the posterior location joint probabilities.
 Each row of the file corresponds to a pair of tiles and a time instant.
}
