% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPostLocProb.R
\name{readPostLocProb}
\alias{readPostLocProb}
\title{Reads a file with the posterior location probabilities.}
\usage{
readPostLocProb(path, prefixName, deviceID)
}
\arguments{
\item{path}{The path to the location where the posterior location probabilities are stored. The file with the
location probabilities should have the name \code{postLocDevice_ID.dt.csv} where \code{ID} is replaced with the
device ID.}

\item{prefixName}{The file name prefix. The whole file name is composed by a concatenation of \code{prefixName},
\code{_},  \code{deviceID} and the extension \code{.dt.csv}.}

\item{deviceID}{The device ID for which the posterior location probabilities are read.}
}
\value{
A data.table object with the posterior location probabilities for the device with ID equals to deviceID. It
  has the following columns: \code{time, tile, prob}. Only the nonzero probabilities are returned.
}
\description{
Reads a .csv file with the posterior location probabilities. There are 3 values on each row of the file:
  \code{time, tile, prob}. Only the nonzero probabilities are given in this file.
}
