% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aprioriOneDevice.R
\name{aprioriOneDeviceProb}
\alias{aprioriOneDeviceProb}
\title{Apriori probability of a person to have one mobile device.}
\usage{
aprioriOneDeviceProb(prob2Devices, ndevices)
}
\arguments{
\item{prob2Devices}{The probability of a person to have 2 devices. In case of using simulated data, this parameter is read
from the simulation configuration file.}

\item{ndevices}{The number of devices detected by the network during the time horizon under consideration.}
}
\description{
Apriori probability of a person to have one mobile device. It is computed simply as (2*Ndev1-Ndev)/Ndev1 where
Ndev1 is the number of devices that are in a 1-to-1 correspondence with the persons that hold them.
}
