% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCentroidProbs.R
\name{buildCentroidProbs}
\alias{buildCentroidProbs}
\title{build a concatenation between centroids and location probabilities}
\usage{
buildCentroidProbs(centroids, postLoc, t)
}
\description{
build a concatenation between centroids and location probabilities
}
