% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDuplicityTable1to1.R
\name{buildDuplicityTable1to1}
\alias{buildDuplicityTable1to1}
\title{Builds a datatable with device IDs and the duplicity probability for each device.}
\usage{
buildDuplicityTable1to1(res, devices, Pii)
}
\arguments{
\item{a}{vector with device IDs.}
}
\value{
a data.table object with two columns: the device IDs and the corresponding duplicity probability for each
  device.
}
\description{
Builds a datatable with with two columns: device IDs and the duplicity probability for each device. This
  function is called in case of using the "1to1" method of computing the dupicity probabilities, receiving a list of
  matrix objects returned by each worker node with lines correponding to a subset of devices. These parts a put
  together and form a symmetic matrix used to compute the duplicity probability. It is a utility function and it is
  not accesible from outside the package.
}
