% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDuplicityTablePairs.R
\name{buildDuplicityTablePairs}
\alias{buildDuplicityTablePairs}
\title{Builds a data.table object with device IDs and the duplicity probability for each device.}
\usage{
buildDuplicityTablePairs(res, devices)
}
\arguments{
\item{res}{A list of data.table objects returned by each worker node with the device IDs on the first two columns and
the corresponding duplicity probability on the third column for (the pair of devices in the first two columns).}

\item{devices}{A vector with device IDs.}
}
\value{
A data.table object with two columns: the device IDs and the corresponding duplicity probability for each
  device.
}
\description{
Builds a data.table object with two columns: device IDs and the duplicity probability for each device.
  This function is called in case of using the "pairs" method of computing the duplicity probabilities, receiving a
  list of data.tables objects returned by each worker node with the device IDs on the first two columns and the
  corresponding duplicity probability on the third column. It is a utility function and it is not accesible from
  outside the package.
}
