% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDuplicityTrajectory.R
\name{computeDuplicityTrajectory}
\alias{computeDuplicityTrajectory}
\title{Computes the duplicity probabilities for each device using the trajectory approach.}
\usage{
computeDuplicityTrajectory(path, devices, gridParams, pairs, P1, T, gamma)
}
\description{
Computes the duplicity probabilities for each device using the trajectory approach.
}
