% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModels.R
\name{fitModels}
\alias{fitModels}
\title{Fits the HMM models for each device.}
\usage{
fitModels(ndevices, model, connections, parallel = TRUE)
}
\arguments{
\item{ndevices}{The number of devices.}

\item{model}{The HMM model returned by \code{getGenericModel()} function. This model is to be fit for each device.}

\item{connections}{A matrix whose elements are the antenna ID where a device is connected at every time instant.
This matrix is returned by \code{getConnections()} function.}

\item{parallel}{If TRUE the computations are done in parallel to reduce the running time using all the available
cores. It creates a cluster of working nodes, splits the devices equally and assign a partition of devices to each
culster. For Unix-like operating systems, this functions uses a "FORK" cluster while for Windows it uses a "SOCK"
cluster.}
}
\value{

}
\description{
Fits the HMM model for each device usin the \code{fit()} function from \code{destim} package.
}
