% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmissionProbs.R
\name{getEmissionProbs}
\alias{getEmissionProbs}
\title{Builds the emission probabilities for the HMM used to estimate the prosterior location probabilitities.}
\usage{
getEmissionProbs(
  nrows,
  ncols,
  signalFileName,
  sigMin,
  handoverType = "strength",
  simulatedData = TRUE
)
}
\value{
Returns a matrix with the emmission probabilitiews for the HMM. The number of rows equals the numberof
tiles in the grid and the number of columns equals the number of antennas.
}
\description{
Builds the emission probabilities needed for the HMM used to estimate the posterior location probabilitities. 
In case of using simulated data, these probabilities are build using the signal strength or signal quality outputed
by the simulation software for each tile in the grid.
}
