% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmissionProbsJointModel.R
\name{getEmissionProbsJointModel}
\alias{getEmissionProbsJointModel}
\title{Builds the emission probabilities for the joint HMM.}
\usage{
getEmissionProbsJointModel(emisionProbs)
}
\value{
Returns a matrix with the joint emission probabilities for the HMM. The number of rows equals the number of
tiles in the grid and the number of columns equals the number of antennas plus one.
}
\description{
Builds the emissions probabilities needed for the joint HMM used to estimate the posterior location probabilitities. 
In case of using simulated data, these probabilities are build using the signal strength or signal quality outputed
by the simulation software for each tile in the grid.
}
