% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeDelta.R
\name{modeDelta}
\alias{modeDelta}
\title{Returns the mode of delta distribution}
\usage{
modeDelta(deltaDistribution)
}
\arguments{
\item{deltaDistribution}{a data.table object that could be deltaX or deltaY distribution. The table has two columns: delta and p.}
}
\value{
the mode of the delta distribution
}
\description{
Returns the mode of the deltaX or deltaY distribution
}
