% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPostLocProb.R
\name{readPostLocProb}
\alias{readPostLocProb}
\title{Reads a file with the posterior location probabilities.}
\usage{
readPostLocProb(path, deviceID)
}
\arguments{
\item{path}{The path to the location where the posterior location probabilities are stored.}

\item{deviceID}{The device ID for which the posterior location probabilities are read.}
}
\value{
a data.table object with the posterior location probabilities for the device with Id
  equals to deviceID. A row corresponds to a tile and a column corresponds to a time instant.
}
\description{
Reads a .csv file with the posterior location probabilities. Each row of the file
  corresponds to a tile and each column corresponds to a time instant.
}
